refactor: pagesに入ってた処理を全てfeaturesに移して、pagesにはそのコンポーネントを呼び出す処理のみにした
